@@ -1,8 +1,8 @@
 import 'dart:async';
 import 'package:flutter/services.dart';
 
-typedef void OnRewardListener(int quantity);
-typedef void SurveyAvailableListener(int survey);
+typedef void OnRewardListener(int? quantity);
+typedef void SurveyAvailableListener(int? survey);
 typedef void RewardCenterOpenedListener();
 typedef void RewardCenterClosedListener();
 
@@ -19,33 +19,33 @@ class RapidoReach {
     _channel.setMethodCallHandler(_platformCallHandler);
   }
 
-  static OnRewardListener _onRewardListener;
-  static SurveyAvailableListener _surveyAvailableListener;
-  static RewardCenterOpenedListener _rewardCenterOpenedListener;
-  static RewardCenterClosedListener _rewardCenterClosedListener;
+  static OnRewardListener? _onRewardListener;
+  static SurveyAvailableListener? _surveyAvailableListener;
+  static RewardCenterOpenedListener? _rewardCenterOpenedListener;
+  static RewardCenterClosedListener? _rewardCenterClosedListener;
 
-  Future<void> init({String apiToken, String userId}) async {
+  Future<void> init({String? apiToken, String? userId}) async {
     assert(apiToken != null && apiToken.isNotEmpty);
     return _channel.invokeMethod(
         "init", <String, dynamic>{"api_token": apiToken, "user_id": userId});
   }
 
-  Future<void> show({String placementID}) {
+  Future<void> show({String? placementID}) {
     return _channel
         .invokeMethod("show", <String, dynamic>{"placementID": placementID});
   }
 
-  Future<void> setNavBarText({String text}) {
+  Future<void> setNavBarText({String? text}) {
     return _channel
         .invokeMethod('setNavBarText', <String, dynamic>{'text': text});
   }
 
-  Future<void> setNavBarColor({String color}) {
+  Future<void> setNavBarColor({String? color}) {
     return _channel
         .invokeMethod('setNavBarColor', <String, dynamic>{'color': color});
   }
 
-  Future<void> setNavBarTextColor({String text_color}) {
+  Future<void> setNavBarTextColor({String? text_color}) {
     return _channel.invokeMethod(
         'setNavBarTextColor', <String, dynamic>{'text_color': text_color});
   }
@@ -56,37 +56,37 @@ class RapidoReach {
 
     switch (call.method) {
       case "onReward":
-        _onRewardListener(call.arguments);
+        _onRewardListener!(call.arguments);
         break;
 
       case "rapidoReachSurveyAvailable":
-        _surveyAvailableListener(call.arguments);
+        _surveyAvailableListener!(call.arguments);
         break;
 
       case "onRewardCenterOpened":
-        _rewardCenterOpenedListener();
+        _rewardCenterOpenedListener!();
         break;
 
       case "onRewardCenterClosed":
-        _rewardCenterClosedListener();
+        _rewardCenterClosedListener!();
         break;
       default:
         print('Unknown method ${call.method}');
     }
   }
 
-  void setOnRewardListener(OnRewardListener onRewardListener) =>
+  void setOnRewardListener(OnRewardListener? onRewardListener) =>
       _onRewardListener = onRewardListener;
 
   void setSurveyAvaiableListener(
-          SurveyAvailableListener surveyAvailableListener) =>
+          SurveyAvailableListener? surveyAvailableListener) =>
       _surveyAvailableListener = surveyAvailableListener;
 
   void setRewardCenterOpened(
-          RewardCenterOpenedListener rewardCenterOpenedListener) =>
+          RewardCenterOpenedListener? rewardCenterOpenedListener) =>
       _rewardCenterOpenedListener = rewardCenterOpenedListener;
 
   void setRewardCenterClosed(
-          RewardCenterClosedListener rewardCenterClosedListener) =>
+          RewardCenterClosedListener? rewardCenterClosedListener) =>
       _rewardCenterClosedListener = rewardCenterClosedListener;
 }
